# docs/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "sc-tournament-system.herokuapp.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
    "/learners/search": {
      "post": {
        "description": "Поиск учеников с параметрами страницы",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "page": {
                  "example": "integer"
                },
                "pageSize": {
                  "example": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/learners": {
      "post": {
        "description": "Добавление нового ученика",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "role_id": {
                  "example": "integer"
                },
                "last_name": {
                  "example": "string"
                },
                "first_name": {
                  "example": "string"
                },
                "second_name": {
                  "example": "string"
                },
                "birthday": {
                  "example": "dateonly"
                },
                "phone_number": {
                  "example": "integer, optional"
                },
                "email": {
                  "example": "example@example.com"
                },
                "password": {
                  "example": "string"
                },
                "city": {
                  "example": "integer"
                },
                "school": {
                  "example": "integer"
                },
                "class": {
                  "example": "integer"
                },
                "fav_sport_types": {
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/learners/{id}": {
      "delete": {
        "description": "Удаление ученика",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "description": "Обновление данных ученика",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "role_id": {
                  "example": "integer"
                },
                "last_name": {
                  "example": "string"
                },
                "first_name": {
                  "example": "string"
                },
                "second_name": {
                  "example": "string"
                },
                "birthday": {
                  "example": "dateonly"
                },
                "phone_number": {
                  "example": "integer"
                },
                "email": {
                  "example": "example@example.com"
                },
                "password": {
                  "example": "string"
                },
                "city": {
                  "example": "integer"
                },
                "school": {
                  "example": "integer"
                },
                "class": {
                  "example": "integer"
                },
                "fav_sport_types": {
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/teachers/search": {
      "post": {
        "description": "Поиск учителей с параметрами страницы",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "page": {
                  "example": "integer"
                },
                "pageSize": {
                  "example": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/teachers": {
      "post": {
        "description": "Добавление нового учителя",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "role_id": {
                  "example": "integer"
                },
                "last_name": {
                  "example": "string"
                },
                "first_name": {
                  "example": "string"
                },
                "second_name": {
                  "example": "string"
                },
                "birthday": {
                  "example": "dateonly"
                },
                "phone_number": {
                  "example": "integer, optional"
                },
                "email": {
                  "example": "example@example.com"
                },
                "password": {
                  "example": "string"
                },
                "city": {
                  "example": "integer"
                },
                "school": {
                  "example": "integer"
                },
                "class": {
                  "example": "integer, optional"
                },
                "access": {
                  "example": "boolean"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/teachers/{id}": {
      "delete": {
        "description": "Удаление ученика",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "description": "Обновление ученика",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "role_id": {
                  "example": "integer"
                },
                "last_name": {
                  "example": "string"
                },
                "first_name": {
                  "example": "string"
                },
                "second_name": {
                  "example": "string"
                },
                "birthday": {
                  "example": "dateonly"
                },
                "phone_number": {
                  "example": "integer"
                },
                "email": {
                  "example": "example@example.com"
                },
                "password": {
                  "example": "string"
                },
                "city": {
                  "example": "integer"
                },
                "school": {
                  "example": "integer"
                },
                "class": {
                  "example": "integer"
                },
                "access": {
                  "example": "boolean"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/tournament/{tournamentId}/match/search": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "tournamentId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "page": {
                  "example": "any"
                },
                "pageSize": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/tournament/{tournamentId}/match": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "tournamentId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/tournament/{tournamentId}/match/{matchId}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "tournamentId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "matchId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "tournamentId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "matchId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/tournament/search": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "page": {
                  "example": "any"
                },
                "size": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/tournament/{id}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/tournament": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "tournament_type_id": {
                  "example": "any"
                },
                "sport_type": {
                  "example": "any"
                },
                "start_date": {
                  "example": "any"
                },
                "end_date": {
                  "example": "any"
                },
                "school_or_city": {
                  "example": "any"
                },
                "team_size": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}